Update dependencies, especially pyjnius
Update minimum versions of all dependencies.

The pyjnius upgrade is especially key because this version (1.2.1) is the first
one with binary wheels on pypi and it also fixes a long-standing string bug
that we had to work around in many places. Unfortunately it also introduces a
bug with visibility of methods on subclasses so we need to add a cast where we
shouldn't need to. Our use of Java isn't heavy enough to make this a problem at
the moment.